+++ packages/larva/__tests__/lib/generateStatic.test.js
@@ -1,14 +1,14 @@
 const path = require( 'path' );
-const fs = require( 'fs' );
 const exec = require( 'child_process' ).exec;
 const app = require( '../../lib/server' );
+const fs = require( 'fs-extra' );
 
 const port = 5555;
 const urlBase = 'http://localhost:' + port + '/larva';
 const generateStatic = require( '../../lib/generateStatic' );
 
 const fixture = path.join( __dirname, '../fixtures' );
-const buildPath = path.join( fixture, './build/html' );
+const buildPath = path.join( fixture, './build/html/project' );
 
 beforeAll( ( done ) => {
 	app.listen( port, () => {
@@ -17,9 +17,9 @@ beforeAll( ( done ) => {
 });
 
 describe( 'generateStatic', () => {
-	
+
 	beforeEach( ( done ) => {
-	
+
 		exec( 'mkdir -p ' + buildPath, ( err ) => {
 			if ( err ) {
 				console.error( err );
@@ -76,9 +76,53 @@ describe( 'generateStatic', () => {
 		}, urlBase );
 	} );
 
+	it( 'copies static assets', ( done ) => {
+		const routesArr = [
+			'components/c-button/brand-basic'
+		];
+
+		generateStatic( routesArr, buildPath, () => {
+
+			// This is kinda sloppy and may be annoying to maintain, and there is
+			// surely a more elegant way to test this functionality than checking
+			// if the file exists...
+
+			// JS
+			expect(
+				fs.existsSync( path.join( buildPath, '../assets/build/js/larva-ui.js' ) )
+			).toBe( true );
+
+			// CSS
+			expect(
+				fs.existsSync( path.join( buildPath, '../assets/build/css/common.inline.css' ) )
+			).toBe( true );
+
+			// Images
+			expect(
+				fs.existsSync( path.join( buildPath, '../assets/build/images/test.png' ) )
+			).toBe( true );
+
+			// SVG
+			expect(
+				fs.existsSync( path.join( buildPath, '../assets/build/svg/defs/sprite.svg' ) )
+			).toBe( true );
+
+			// Public dir
+			expect(
+				fs.existsSync( path.join( buildPath, '../assets/public/some-library.js' ) )
+			).toBe( true );
+
+			done();
+
+		}, urlBase );
+
+	} );
+
 	afterEach( ( done ) => {
 
-		exec( 'rm -r ' + buildPath, ( err ) => {
+		const htmlDir = path.join( buildPath, '../' );
+
+		exec( 'rm -r ' + htmlDir, ( err ) => {
 			if ( err ) {
 				console.error( err );
 			}
@@ -86,7 +130,7 @@ describe( 'generateStatic', () => {
 		} );
 
 	} );
-	
+
 } );
 
 afterAll( async ( done ) => {